pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphsql"
version = "0.1.0"
description = "Automatic REST and GraphQL API for arbitrary databases"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Apache-2.0" }
authors = [
    { name = "Anton Feldmann", email = "anton.feldmann@gmail.com" }
]
keywords = ["fastapi", "graphql", "sqlalchemy", "rest", "api", "automatic", "generator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.20.0",
    "sqlalchemy>=1.4.0",
    "strawberry-graphql[fastapi]>=0.200.0",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.0",  # PostgreSQL
    "pymysql>=1.0.0",  # MySQL
    "cryptography>=38.0.0",  # for MySQL
    "pydantic>=1.10.0",
    "loguru>=0.6.0",
    "typer[all]>=0.12.0",  # CLI
    "rich>=13.0.0",  # Rich formatting
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.0.250",
    "mypy>=1.0.0",
    "sphinx>=6.0.0",
    "sphinx-autodoc-typehints>=1.20.0",
    "behave>=1.2.6",
]
sqlite = [
    "aiosqlite>=0.19.0",
]

postgres = [
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
]

mysql = [
    "pymysql>=1.1.0",
    "aiomysql>=0.2.0",
]
cloud = [
    "sqlalchemy-hana>=0.13.0",
    "hdbcli>=2.21.0",
    "sqlalchemy-redshift>=0.8.14",
    "snowflake-sqlalchemy>=1.5.0",
]
all = [
    "graphsql[sqlite,postgres,mysql,cloud]",
]


[project.urls]
Homepage = "https://github.com/afeldman/graphsql"
Documentation = "https://github.com/afeldman/graphsql#readme"
Repository = "https://github.com/afeldman/graphsql"
Issues = "https://github.com/afeldman/graphsql/issues"

[project.scripts]
graphsql = "graphsql.cli:main"
graphsql-server = "graphsql.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/graphsql"]

[tool.hatch.version]
path = "src/graphsql/__init__.py"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = "strawberry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=graphsql",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
